Add more `dict`-like methods for `PickleInterface`. Re-arrange `main.py`'s bot config
